@@ -1,12 +1,13 @@
 import React, { Component } from 'react'
 import { Link } from 'react-router-dom'
 import { withAuth } from "../lib/AuthProvider";
-const axios = require('axios');
+import recipe from '../lib/recipe-service';
+
 class Search extends Component {
   constructor(props) {
-    const pantryContents = (pantryItems) => { 
-      return pantryItems.map((item) => item.item )
-    } 
+    const pantryContents = (pantryItems) => {
+      return pantryItems.map((item) => item.item)
+    }
     super(props)
     this.state = {
       searchTerm: '',
@@ -15,24 +16,14 @@ class Search extends Component {
     }
   }
 
-  handleSearch = (e) => {  
-    const {name, value} = e.target;
-    this.setState({[name]: value});
-    axios.get(`http://localhost:5000/recipes/search?s=${e.target.value}`)
-      .then(({ data }) => {
-        this.setState({recipes: data})
+  handleSearch = (e) => {
+    const { name, value } = e.target;
+    this.setState({ [name]: value });
+    recipe.search(e.target.value)
+      .then((data) => {
+        this.setState({ recipes: data })
       })
       .catch((error) => console.log(error));
-  }
-
-  handleLucky = (e) => {
-    e.preventDefault();
-    const searchForItems = this.state.pantry;
-    axios.post('http://localhost:5000/recipes/search', {
-      searchForItems
-    })
-    .then((response) => console.log(response))
-    .catch((error) => console.log(error));
   }
 
   render() {
@@ -41,11 +32,7 @@ class Search extends Component {
         <form>
           <input type="text" autoFocus name="searchTerm" value={this.state.searchTerm} placeholder="Find a recipe" onChange={(e) => this.handleSearch(e)} />
         </form>
-        {this.props.isLoggedin ?
-        <button onClick={(e) => this.handleLucky(e)}>I'm feeling lucky</button>
-        :
-        null }
-        {this.state.recipes.map((recipe) => 
+        {this.state.recipes.map((recipe) =>
           <Link key={recipe._id} to={{
             pathname: `/recipes/${recipe._id}`,
             state: { selectedRecipe: recipe }
